requirements.txt
telethon==1.38.1
python-dotenv==1.0.1
ghkit==0.0.3
schedule==1.2.2
fastapi==0.109.1
uvicorn==0.30.1
python-socketio==5.11.4
fastapi-socketio==0.0.10
passlib==1.7.4
tortoise-orm==0.21.5
httpx~=0.27.2
redis==5.2.1
aiomysql==0.2.0
aerich==0.7.2
python-jose==3.3.0
python-multipart==0.0.19